Version 1.0.5

+++ calc/src/main/java/edu/usmp/calc/model/Result.java
@@ -2,27 +2,27 @@ package edu.usmp.calc.model;
 
 public class Result {
 
-	private long n1; 
-	private long n2;
-	private long result;
+	private double n1; 
+	private double n2;
+	private double result;
 	private char op;
 	
-	public long getN1() {
+	public double getN1() {
 		return n1;
 	}
-	public void setN1(long n1) {
+	public void setN1(double n1) {
 		this.n1 = n1;
 	}
-	public long getN2() {
+	public double getN2() {
 		return n2;
 	}
-	public void setN2(long n2) {
+	public void setN2(double n2) {
 		this.n2 = n2;
 	}
-	public long getResult() {
+	public double getResult() {
 		return result;
 	}
-	public void setResult(long result) {
+	public void setResult(double result) {
 		this.result = result;
 	}
 	public char getOp() {
